Change id type from int to long in Film and User

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -16,7 +16,7 @@ import java.util.Set;
 @Data
 @Builder
 public class Film {
-    private int id;
+    private long id;
     @NotBlank
     private String name;
     @Size(max = 200)
@@ -27,5 +27,5 @@ public class Film {
     private LocalDate releaseDate;
     @Positive
     private int duration;
-    private final Set<Integer> likes = new HashSet<>();
+    private final Set<Long> likes = new HashSet<>();
 }
